MassSpecWavelet-package {MassSpecWavelet}	R Documentation

Peak detection of mass spectrum by Wavelet transform based methods

Description

Process Mass Spectrum (MS) by Wavelet Transforms-based algorithms

Details

Package: 	 MassSpecWavelet

Type: 	 Package

Version: 	 1.0.4

Date: 	 2007-04-05

License: 	 GPL 2 or newer

MassSpecWavelet R package is aimed to process Mass Spectrometry (MS) data mainly based on Wavelet Transforms. The current version only supports the peak detection based on Continuous Wavelet Transform (CWT). Future versions will include more functions covering entire MS data processes.

Author(s)

Pan Du, Simon Lin

Maintainer: Pan Du <dupan@northwestern.edu>

References

Du, P., Kibbe, W.A. and Lin, S.M. (2006) Improved peak detection in mass spectrum by incorporating continuous wavelet transform-based pattern matching, Bioinformatics, 22, 2059-2065.

Examples

data(exampleMS)
SNR.Th <- 3
peakInfo <- peakDetectionCWT(exampleMS, SNR.Th=SNR.Th)
majorPeakInfo = peakInfo$majorPeakInfo
peakIndex <- majorPeakInfo$peakIndex
plotPeak(exampleMS, peakIndex, main=paste('Identified peaks with SNR >', SNR.Th)) 

[Package MassSpecWavelet version 1.52.0 Index]