extendLength {MassSpecWavelet}	R Documentation

Extend the length of a signal or matrix

Description

Extend the length of a signal or matrix by row

Usage

extendLength(x, addLength = NULL, method = c("reflection", "open", "circular"), direction = c("right", "left", "both"))

Arguments

x
a vector or matrix with column with each column as a signal

addLength
the length to be extended

method
three methods available, c("reflection", "open", "circular"). By default, it is "reflection".

direction
three options available: c("right", "left", "both")

Value

return the extended vector or matrix.

Author(s)

Pan Du

See Also

extendNBase

Examples

	# a = matrix(rnorm(9), 3)
	# extendLength(a, 3, direction='right') 	## not exposed function

[Package MassSpecWavelet version 1.52.0 Index]