extendNBase {MassSpecWavelet}	R Documentation

Extend the row number of a matrix as the exponential of base N

Description

Extend the data as the exponential of base N by increasing row number.

Usage

extendNBase(x, nLevel=1, base=2, ...) 

Arguments

x
data matrix

nLevel
the level of DWT decomposition. Basically, it is equivalent to changing the 'base' as base\^nLevel

base
the base, 2 by default

...
other parameters of used by extendLength

Details

The method 'open' is padding the the matrix with the last row.

Value

Return a extended matrix

Author(s)

Pan Du

See Also

extendLength

Examples

	# a = matrix(rnorm(9), 3)
	# extendNBase(a)	## not exposed function

[Package MassSpecWavelet version 1.52.0 Index]