getLocalMaximumCWT {MassSpecWavelet}	R Documentation

Identify the local maximum of each column in 2-D CWT coefficients matrix

Description

Identify the local maximum of each column in 2-D CWT coefficients matrix by using a slide window. The size of slide window linearly changes from the coarse scale (bigger window size) to detail scale. The scale of CWT increases with the column index.

Usage

getLocalMaximumCWT(wCoefs, minWinSize= 5, amp.Th = 0)

Arguments

wCoefs
2-D CWT coefficients, each column corresponding to CWT coefficient at one scale. The column name is the scale.

minWinSize
The minimum slide window size used.

amp.Th
The minimum peak amplitude.

Value

return a matrix with same dimension as CWT coefficient matrix, wCoefs. The local maxima are marked as 1, others are 0.

Author(s)

Pan Du

See Also

localMaximum

Examples

	data(exampleMS)
	scales <- seq(1, 64, 3)
	wCoefs <- cwt(exampleMS[5000:11000], scales=scales, wavelet='mexh')
	
	localMax <- getLocalMaximumCWT(wCoefs)
	plotLocalMax(localMax)

[Package MassSpecWavelet version 1.52.0 Index]