getRidgeValue {MassSpecWavelet}	R Documentation

Get the CWT coefficient values corresponding to the peak ridge

Description

Get the CWT coefficient values corresponding to the peak ridge

Usage

getRidgeValue(ridgeList, wCoefs, skip = 0)

Arguments

ridgeList
a list of ridge lines

wCoefs
2-D CWT coefficients

skip
the CWT scale level to be skipped, by default the 0 scale level (raw spectrum) is skipped.

Value

A list of ridge values corresponding to the input ridgeList.

Author(s)

Pan Du

[Package MassSpecWavelet version 1.52.0 Index]